rename oidc strategy file
